ai gen code: I need a gps tracking  app that will be executed in the android phone in kiosk mode, I will connect the phone to the computer and set device owner. So the app can have super privileges and access to special apis
This app will be the only application running on the phone, The APP will be idle state most of the time. I need to define the timeslots where the application wont be idle state and will be awake statw.

When awake, it will turn on the network data, get the gps position and send the location to a mqtt endpoint using wss, every 30 secs.
Once the awake timeslot is finished, need to go back to idle state and  disable gps, data and set airplane mode.

The awake timeslots can be hardcoded somewhere in the code in  alist

example:
8:00 -> 9:00
14:00 -> 15:00
18:00 -> 19:00

make sure that when the device is idle, all is disabled to save on battery.

Maybe show a simple UI for debugging, displaying the current state, and if awake state, show mqtt connection details, and every time that the gps location is acquired and sent.

Once this is working, create a detailed steps to enrol device as device owner, and set it in kiosk mode and the set the app to be the only one running on the phone.

Also programmatically make sure that all packages don't have access to the network data and only my app  can have access to the network data. also any other system app that is required to send data through my app. the idea is to use little data, as I am charged for every kb sent, so I need to makesure my app is the only one accessing the network when in the wawake state.

Also I need a new state, that wakes the phone every hour and report the battery level in the UI, and if battery below a limit it sends another message in the mqtt in a different topic, when in this state try to use wifi first if not available then use network data to send the battery warning message.